поправил тесты
