+++ src/reducer/data-reducer.test.js
@@ -4,8 +4,7 @@ import {randomOffers as offers} from "../mocks/offers";
 describe(`Data reducer is correct`, () => {
   it(`returns initial state without parameters`, () => {
     expect(dataReducer(undefined, {})).toEqual({
-      cityName: `Amsterdam`,
-      offers: []
+      cityName: `Amsterdam`
     });
   });
   it(`correctly change city`, () => {
@@ -25,48 +24,6 @@ describe(`Data reducer is correct`, () => {
       offers
     });
   });
-  it(`correctly change offers`, () => {
-    const state = {
-      cityName: `Amsterdam`,
-      offers
-    };
-    const offersMock = [
-      {
-        name: `Beautiful & luxurious apartment at great location`,
-        image: `img/apartment-01.jpg`,
-        price: 120,
-        rating: 5,
-        isPremium: false,
-        isFavorite: true,
-        coordinates: [52.3909553943508, 4.85309666406198],
-        city: `Amsterdam`,
-        id: 0
-      },
-      {
-        name: `Wood and stone place`,
-        image: `img/room.jpg`,
-        price: 80,
-        rating: 10,
-        isPremium: true,
-        isFavorite: false,
-        coordinates: [52.369553943508, 4.85309666406198],
-        city: `Amsterdam`,
-        id: 1
-      }
-    ];
-
-    const offersChangeAction = {
-      type: `CHANGE_OFFERS`,
-      payload: {
-        offers: offersMock
-      }
-    };
-    expect(dataReducer(state, offersChangeAction)).toEqual({
-      cityName: `Amsterdam`,
-      offers: offersMock
-    });
-  });
-
   it(`returns state on incorrect action`, () => {
     const offersMock = [
       {
