+++ src/reducer/selectors.js
@@ -4,7 +4,6 @@ const data = NameSpace.DATA;
 const fetchOffers = NameSpace.FETCH_OFFERS;
 
 export const getOffers = (state) => {
-  console.log(state);
   return state[fetchOffers].offers;
 };
 
@@ -14,14 +13,6 @@ export const getCityName = (state) => {
 
 export const getOffersIsLoading = (state) => {
   return state[fetchOffers].isLoading;
-};
-
-export const getOffersIsLoaded = (state) => {
-  return state[fetchOffers].isLoaded;
-};
-
-export const getOffersIsFailed = (state) => {
-  return state[fetchOffers].isFailed;
 };
 
 export const getOffersLoadError = (state) => {
